[Test] Split ops test for selected way
