+++ tests/e2e/pull_request/light/four_card/test_qwen3_5.py
