+++ tests/e2e/pull_request/light/one_card/test_vlm.py
@@ -24,11 +24,16 @@ Run `pytest tests/test_offline_inference.py`.
 import os
 from unittest.mock import patch
 
+import pytest
 from vllm import SamplingParams
 from vllm.assets.audio import AudioAsset
 from vllm.assets.image import ImageAsset
 
 from tests.e2e.conftest import VllmRunner
+
+WHISPER_MODELS = [
+    "openai-mirror/whisper-large-v3-turbo",
+]
 
 
 @patch.dict(os.environ, {"VLLM_WORKER_MULTIPROC_METHOD": "spawn"})
@@ -113,3 +118,20 @@ def test_multimodal_audio():
 
         assert outputs is not None, "Generated outputs should not be None."
         assert len(outputs) > 0, "Generated outputs should not be empty."
+
+
+@pytest.mark.parametrize("model", WHISPER_MODELS)
+@patch.dict(os.environ, {"VLLM_WORKER_MULTIPROC_METHOD": "spawn"})
+def test_whisper(model) -> None:
+    prompts = ["<|startoftranscript|><|en|><|transcribe|><|notimestamps|>"]
+    audios = [AudioAsset("mary_had_lamb").audio_and_sample_rate]
+
+    sampling_params = SamplingParams(temperature=0.2, max_tokens=10, stop_token_ids=None)
+
+    with VllmRunner(
+        model, max_model_len=448, max_num_seqs=5, dtype="bfloat16", block_size=128, gpu_memory_utilization=0.9
+    ) as runner:
+        outputs = runner.generate(prompts=prompts, audios=audios, sampling_params=sampling_params)
+
+    assert outputs is not None, "Generated outputs should not be None."
+    assert len(outputs) > 0, "Generated outputs should not be empty."
